+++ docs/changelog.rst
@@ -5,7 +5,10 @@ CHANGELOG
 2.4.4.dev0
 ------------------
 
+**Bug fixes**
+
 * Show pending import/sync tasks
+* Add a command to clean orphaned attached media
 
 
 2.4.3 (2015-11-27)